app/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic[email]
mysql-connector-python
python-dotenv
